requirements.txt
pytest-playwright
pandas
